+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   within the annotation driven component model.
 * **\[gwt-output-qa\]** Upgrade the asserts to verify that the class `arez.component.ComponentState`
   is optimized out in production builds.
+* **\[core\]** Introduce a helper method `arez.ArezTestUtil.resetState()` responsible for resetting
+  context and zone state. This is occasionally required in tests.
 
 ### [v0.92](https://github.com/arez/arez/tree/v0.92) (2018-06-17)
 [Full Changelog](https://github.com/arez/arez/compare/v0.91...v0.92)